RUM-1153: Migrate objc interfaces to Core module
