+++ DatadogCore/Tests/DatadogObjc/ObjcAPITests/DDURLSessionInstrumentationTests+apiTests.m
@@ -7,6 +7,8 @@
 #import <XCTest/XCTest.h>
 #include <sys/wait.h>
 @import DatadogObjc;
+@import DatadogCore;
+@import DatadogInternal;
 @import DatadogTrace;
 
 #import <Foundation/Foundation.h>
